fix(scan-matching): desabilitar dθ + corrigir float32 + gain 0.7
Diagnóstico do teste (20 correções em 23s):
- Score 0.31 foi rejeitado indevidamente: xy_offsets em float32 produzia
  -0.20000003 > max_correction_m=0.20 → rejeição falsa. Corrigido para float64.
- dθ acumulou +13° em 23s (ruído em mapa com 4.7% ocupado): DESABILITADO
  (SCAN_MATCH_THETA_RANGE_DEG=0.0). Ângulo fica com odometria+CTE.
- Fator de amortecimento SCAN_MATCH_CORRECTION_GAIN=0.7 aplicado em dx/dy
  para reduzir impacto de correções ruidosas.
- Benefício velocidade: 9×9×1=81 candidatos (era 9×9×7=567) → ~50ms esperado.

Made-with: Cursor

+++ src/core/lidar_pose_corrector.py
@@ -331,9 +331,10 @@ class LidarPoseCorrector:
             dists_valid = dists_valid[::step]
 
         # ── Grade de candidatos ──────────────────────────────────────────────
+        # float64 evita erros de precisão nos limites (ex: -0.20 armazenado como -0.2000003 em float32)
         xy_offsets        = np.arange(-self.xy_range_m,
                                        self.xy_range_m + 1e-6,
-                                       self.xy_step_m, dtype=np.float32)
+                                       self.xy_step_m)
         theta_offsets_deg = np.arange(-self.theta_range_deg,
                                        self.theta_range_deg + 1e-6,
                                        self.theta_step_deg)
